Move SetRoomCameraPreferencesEvent to Preferences
